Actually define lighting settings if present
Previously the Parser would output the lighting settings to config.h commented out.

This change makes it so the necessary defines are not commented out, if the imported JSON file has the necessary configuration.

BACKLIGHT_BREATHING is still commented out. Backlight breathing being supported by the hardware is dependent on data that could possibly be inferred from the JSON file, but I'm not currently checking it and haven't looked into it.